Allow users to get route names.

This function allows users to get route names based on the specified bus number. If bus number is not specified, it will provide the all bus number with its corresponding route names.

get_route(bus = NULL, key = Sys.getenv("BUS_CLIENT_KEY"))

Arguments

bus
Bus number.

key
Key for accessing the API.

Value

Dataframe with bus number and its corresponding route name.

Examples

get_route()

#>     bus_number                      route_name
#> 1            1       Bronzeville/Union Station
#> 2            2               Hyde Park Express
#> 3            3                      King Drive
#> 4            4                   Cottage Grove
#> 5            5           South Shore Night Bus
#> 6            6            Jackson Park Express
#> 7            7                        Harrison
#> 8            8                         Halsted
#> 9           8A                   South Halsted
#> 10           9                         Ashland
#> 11          X9                 Ashland Express
#> 12          11                         Lincoln
#> 13          12                       Roosevelt
#> 14         J14                    Jeffery Jump
#> 15          15                   Jeffery Local
#> 16          18                       16th/18th
#> 17          19           United Center Express
#> 18          20                         Madison
#> 19          21                          Cermak
#> 20          22                           Clark
#> 21          24                       Wentworth
#> 22          26             South Shore Express
#> 23          28                    Stony Island
#> 24          29                           State
#> 25          30                   South Chicago
#> 26          31                            31st
#> 27          34                  South Michigan
#> 28          35                       31st/35th
#> 29          36                        Broadway
#> 30          37                        Sedgwick
#> 31          39                        Pershing
#> 32          43                            43rd
#> 33          44                  Wallace-Racine
#> 34          47                            47th
#> 35          48                     South Damen
#> 36          49                         Western
#> 37         49B                   North Western
#> 38         X49                 Western Express
#> 39          50                           Damen
#> 40          51                            51st
#> 41          52               Kedzie/California
#> 42         52A                    South Kedzie
#> 43          53                         Pulaski
#> 44         53A                   South Pulaski
#> 45          54                          Cicero
#> 46         54A       North Cicero/Skokie Blvd.
#> 47         54B                    South Cicero
#> 48          55                        Garfield
#> 49         55A                     55th/Austin
#> 50         55N               55th/Narragansett
#> 51          56                       Milwaukee
#> 52          57                         Laramie
#> 53          59                       59th/61st
#> 54          60                Blue Island/26th
#> 55          62                          Archer
#> 56         62H                   Archer/Harlem
#> 57          63                            63rd
#> 58         63W                       West 63rd
#> 59          65                           Grand
#> 60          66                         Chicago
#> 61          67                  67th-69th-71st
#> 62          68               Northwest Highway
#> 63          70                        Division
#> 64          71                71st/South Shore
#> 65          72                           North
#> 66          73                        Armitage
#> 67          74                       Fullerton
#> 68          75                       74th-75th
#> 69          76                        Diversey
#> 70          77                         Belmont
#> 71          78                        Montrose
#> 72          79                            79th
#> 73          80                     Irving Park
#> 74          81                        Lawrence
#> 75         81W                   West Lawrence
#> 76          82                   Kimball-Homan
#> 77          84                        Peterson
#> 78          85                         Central
#> 79         85A                   North Central
#> 80          86          Narragansett/Ridgeland
#> 81          87                            87th
#> 82          88                         Higgins
#> 83          90                          Harlem
#> 84          91                          Austin
#> 85          92                          Foster
#> 86          93                California/Dodge
#> 87          94                South California
#> 88          95                            95th
#> 89          96                            Lunt
#> 90          97                          Skokie
#> 91         X98                    Avon Express
#> 92         100           Jeffery Manor Express
#> 93         103                      West 103rd
#> 94         106                      East 103rd
#> 95         108                    Halsted/95th
#> 96         111                111th/King Drive
#> 97        111A                 Pullman Shuttle
#> 98         112                 Vincennes/111th
#> 99         115                   Pullman/115th
#> 100        119                  Michigan/119th
#> 101        120   Ogilvie/Streeterville Express
#> 102        121     Union/Streeterville Express
#> 103        124                       Navy Pier
#> 104        125             Water Tower Express
#> 105        126                         Jackson
#> 106        132            Goose Island Express
#> 107        134        Stockton/LaSalle Express
#> 108        135       Clarendon/LaSalle Express
#> 109        136        Sheridan/LaSalle Express
#> 110        143       Stockton/Michigan Express
#> 111        146    Inner Drive/Michigan Express
#> 112        147             Outer Drive Express
#> 113        148      Clarendon/Michigan Express
#> 114        151                        Sheridan
#> 115        152                         Addison
#> 116        155                           Devon
#> 117        156                         LaSalle
#> 118        157            Streeterville/Taylor
#> 119        165                       West 65th
#> 120        169                69th-UPS Express
#> 121        171         U. of Chicago/Hyde Park
#> 122        172           U. of Chicago/Kenwood
#> 123        192 U. of Chicago Hospitals Express
#> 124        201                   Central/Ridge
#> 125        206             Evanston Circulator

get_route(bus = 5)

#>   bus_number            route_name
#> 1          5 South Shore Night Bus